requirements.txt
PyQt5
numpy
scipy
tensorflow
scikit-learn
pyserial
pygame 